% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anesthesia}
\alias{anesthesia}
\title{Anesthetist ratings for paitent suitability for surgery}
\format{A data frame with 315 rows and 3 variables:
\describe{
  \item{ii}{item index, which item is being rated}
  \item{jj}{rater index, which rater is doing the rater}
  \item{y}{what rating was given}
}}
\usage{
anesthesia
}
\description{
A dataset containing anesthetist ratings of paitent readiness for surgery for
5 anesthetists on 45 patients with 4 categories in 'long' format. Each row
represents one unique rating or annotation
}
\keyword{datasets}
