% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{as_mcmc.list}
\alias{as_mcmc.list}
\title{Convert a rater_fit object to a {coda} \code{mcmc.list} object.}
\usage{
as_mcmc.list(fit)
}
\arguments{
\item{fit}{A rater_fit object.}
}
\value{
A {coda} mcmc.list object.
}
\description{
Convert a rater_fit object to a {coda} \code{mcmc.list} object.
}
\examples{
\donttest{

# Fit a model using MCMC (the default).
mcmc_fit <- rater(anesthesia, "dawid_skene")

# Convert it to an mcmc.list
rater_mcmc_list <- as_mcmc.list(mcmc_fit)

}

}
