% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caries}
\alias{caries}
\title{Caries ratings of whether caries are 'sound' or not (c...)}
\format{
A matrix
\describe{
  \item{ii}{item index, which item is being rated}
  \item{jj}{rater index, which rater is doing the rater}
  \item{y}{what rating was given}
}
}
\usage{
caries
}
\description{
A dataset containing anesthetist ratings of patient readiness for surgery for
5 anesthetists on 45 patients with 4 categories in 'long' format. Each row
represents one unique rating or annotation
}
\keyword{datasets}
