% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_convergence}
\alias{check_convergence}
\title{That there has been no divergent transitions or poor convergence}
\usage{
check_convergence(draws)
}
\arguments{
\item{draws}{stanfit object}
}
\description{
That there has been no divergent transitions or poor convergence
}
