% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{check_K}
\alias{check_K}
\title{Helper to check if the prior parameters and data have consistent dimensions}
\usage{
check_K(stan_data, model)
}
\arguments{
\item{model}{the rater_model}

\item{K}{the number of categories}
}
\value{
the fully reliased prior parameters
}
\description{
Helper to check if the prior parameters and data have consistent dimensions
}
