% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{class_conditional_dawid_skene_inits}
\alias{class_conditional_dawid_skene_inits}
\title{Creates inits for the class conditional dawid and skene model}
\usage{
class_conditional_dawid_skene_inits(K, J)
}
\arguments{
\item{K}{number of categories}

\item{J}{number of raters}
}
\value{
inits in the format required by stan
}
\description{
Creates inits for the class conditional dawid and skene model
}
