% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{class_probabilities}
\alias{class_probabilities}
\alias{class_probabilities.mcmc_fit}
\alias{class_probabilities.optim_fit}
\title{Extract latent class probabilites from an object.}
\usage{
class_probabilities(fit, ...)

\method{class_probabilities}{mcmc_fit}(fit, ...)

\method{class_probabilities}{optim_fit}(fit, ...)
}
\arguments{
\item{fit}{A rater fit object.}

\item{...}{Extra arguments.}
}
\value{
A I * K matrix where each element is the probabily of item i being
of class k.
}
\description{
Extract latent class probabilites from an object.
}
