% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley.R
\name{colley}
\alias{colley}
\title{Compute ratings from Colley matrix and net win differential}
\usage{
colley(colley_mtx, net_rec)
}
\arguments{
\item{colley_mtx}{a Colley matrix containing the games played data}

\item{net_rec}{a vector containing the net record of each team}
}
\value{
\code{colley} returns a vector with the Colley rating for
each team.
}
\description{
This function computes the Colley ratings for a league from a Colley matrix
and the net number of wins by each team
}
