% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankplot.R
\name{combine_rankgraphs}
\alias{combine_rankgraphs}
\title{Create graph object for two distinct graph objects}
\usage{
combine_rankgraphs(left, right)
}
\arguments{
\item{x, y}{ranking vectors of comparable items}

\item{other_rankings}{vector of items that might not be in `x` or `y`}
}
\description{
This takes to graph objects created by \code{rankgraph} and positions them
side by side.  To render the graph, use \code{DiagrammeR::render_graph}.
}
