% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{creat_inits}
\alias{creat_inits}
\title{Creates inits for the stan MCMC chains}
\usage{
creat_inits(model, stan_data)
}
\arguments{
\item{model}{rater model}

\item{stan_data}{data in list form}
}
\value{
intits in the format required by stan
}
\description{
Creates inits for the stan MCMC chains
}
