% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{default_alpha}
\alias{default_alpha}
\title{Creates an alpha according to the default prior specification}
\usage{
default_alpha(K)
}
\arguments{
\item{K}{number of categories in the data}
}
\description{
Creates an alpha according to the default prior specification
}
