% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_class.R
\name{extract.fit}
\alias{extract.fit}
\title{Extract method for a rater fit object}
\usage{
\method{extract}{fit}(fit, param = c("theta", "raters", "pi",
  "prevalance", "z", "latent_class"), which = NULL)
}
\arguments{
\item{fit}{a fit object}

\item{which}{Which of the raters' matrices to extract (only used for type =
theta or raters)}

\item{type}{what sort of plot should be plotted}
}
\description{
Extract different paramter estimates from a fit object
}
