% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{extract_latent_class}
\alias{extract_latent_class}
\title{Generic to extract (latent class) from a fit}
\usage{
extract_latent_class(fit, ...)
}
\arguments{
\item{fit}{an object}

\item{...}{extra stuff}
}
\description{
Generic to extract (latent class) from a fit
}
