% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_pi.optim_fit}
\alias{extract_pi.optim_fit}
\title{Extract prevalance information from optim fit object}
\usage{
\method{extract_pi}{optim_fit}(fit, ...)
}
\arguments{
\item{fit}{fit object}

\item{...}{extra args}
}
\description{
Extract prevalence/pi estiamtes from a fit object
}
