% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{extract_pi}
\alias{extract_pi}
\title{Generic to extract pi from a fit}
\usage{
extract_pi(fit, ...)
}
\arguments{
\item{fit}{an object}

\item{...}{extra stuff}
}
\description{
Generic to extract pi from a fit
}
