% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_prevalance}
\alias{extract_prevalance}
\title{Extract prevalance information}
\usage{
extract_prevalance(fit)
}
\arguments{
\item{fit}{fit object}
}
\value{
data.frame of estimated mean category probaility and standard
  deviation
}
\description{
Extract prevalence/pi estiamtes from a fit object
}
