% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{extract_prevalence}
\alias{extract_prevalence}
\title{Function to extract prevalence from a fit}
\usage{
extract_prevalence(fit, ...)
}
\arguments{
\item{fit}{an object}

\item{...}{extra stuff}
}
\description{
Function to extract prevalence from a fit
}
