% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_raters_ds}
\alias{extract_raters_ds}
\title{Extract rater accuarcy estimates for the Dawid Skene models}
\usage{
extract_raters_ds(fit, which = NULL)
}
\arguments{
\item{fit}{fit object}

\item{which}{which raters to extract}
}
\value{
list of matrices containing proabaility confusion matrices for each
 rater
}
\description{
Extract rater accuarcy/theta estimates from a Dawid Skene fit object
}
