% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_raters_multi}
\alias{extract_raters_multi}
\title{Extract rater accuarcy estimates from a multinomial model}
\usage{
extract_raters_multi(fit)
}
\arguments{
\item{fit}{fit object}
}
\value{
list of matrices containing probability confusion matrices for each
 rater
}
\description{
Extract rater accuarcy/theta estimates from a multinomial fit object
}
