% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_theta.mcmc_fit}
\alias{extract_theta.mcmc_fit}
\title{Extract rater accuracy estimates for the Dawid Skene models}
\usage{
\method{extract_theta}{mcmc_fit}(fit, which = NULL, ...)
}
\arguments{
\item{fit}{fit object}

\item{which}{which raters to extract}

\item{...}{extra args}
}
\value{
list of matrices containing probability confusion matrices for each
 rater
}
\description{
Extract rater accuracy/theta estimates from a Dawid Skene fit object
}
