% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_theta.optim_fit}
\alias{extract_theta.optim_fit}
\title{Extract theta parameter for the optim_fit class}
\usage{
\method{extract_theta}{optim_fit}(fit, which = NULL, ...)
}
\arguments{
\item{fit}{fit object}

\item{which}{which raters to extract}

\item{...}{extra args}
}
\value{
array of rater error distibutions
}
\description{
Extract theta parameter for the optim_fit class
}
