% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{extract_theta}
\alias{extract_theta}
\title{Generic to extract theta from a fit}
\usage{
extract_theta(fit, which = NULL, ...)
}
\arguments{
\item{fit}{an object}

\item{which}{which rater to select}

\item{...}{extra args}
}
\description{
Generic to extract theta from a fit
}
