% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_z.mcmc_fit}
\alias{extract_z.mcmc_fit}
\title{Extract latent class estimates from a fit}
\usage{
\method{extract_z}{mcmc_fit}(fit, ...)
}
\arguments{
\item{fit}{fit object}

\item{...}{extra args}
}
\value{
Probalistic latent class measurements
}
\description{
Extract latent class estimates from a fit
}
