% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_z.optim_fit}
\alias{extract_z.optim_fit}
\title{Extract latent class estimates from a optim fit}
\usage{
\method{extract_z}{optim_fit}(fit, ...)
}
\arguments{
\item{fit}{fit object}

\item{...}{extra args}
}
\value{
Probalistic latent class measurements
}
\description{
Extract latent class estimates from a optim fit
}
