% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_class.R
\name{fit}
\alias{fit}
\title{make a rater fit object}
\usage{
fit(model, draws, data)
}
\arguments{
\item{model}{a rater model}

\item{draws}{a stanfit object}

\item{data}{the data used to fit the model}
}
\value{
a rater fit object
}
\description{
make a rater fit object
}
