% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall.R
\name{get_pairs}
\alias{get_pairs}
\title{Auxiliary function, not for export
Returns a list of 2-vectors for vectors x, y:
(x_i, x_j) for i < j and (x_k, y_l) for all k, l
In the ranking interpretation x is the ranking list and y is the list of
items that is not ranked by x.}
\usage{
get_pairs(x, y)
}
\description{
Auxiliary function, not for export
Returns a list of 2-vectors for vectors x, y:
(x_i, x_j) for i < j and (x_k, y_l) for all k, l
In the ranking interpretation x is the ranking list and y is the list of
items that is not ranked by x.
}
