% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{get_stan_file}
\alias{get_stan_file}
\title{Helper get the correct stan file to run for model/data combination}
\usage{
get_stan_file(data, model)
}
\arguments{
\item{data}{a rater_data object}

\item{model}{a rater_model object}
}
\value{
the name (no .stan) of the stan file that should be run
}
\description{
Helper get the correct stan file to run for model/data combination
}
