% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{get_stanfit}
\alias{get_stanfit}
\title{Get the underlying \code{stanfit} object from a \code{rater_fit} object.}
\usage{
get_stanfit(fit)
}
\arguments{
\item{fit}{A \code{rater_fit} object.}
}
\value{
A \code{stanfit} object from rstan.
}
\description{
Get the underlying \code{stanfit} object from a \code{rater_fit} object.
}
\examples{

\donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit

}

}
