% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{is.dawid_skene}
\alias{is.dawid_skene}
\title{Check if model is of type Dawid and skene}
\usage{
is.dawid_skene(model)
}
\arguments{
\item{model}{object of type model}
}
\description{
Check if model is of type Dawid and skene
}
