% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{is.hier_dawid_skene}
\alias{is.hier_dawid_skene}
\title{Check if model is of type Hierarchical Dawid and skene}
\usage{
is.hier_dawid_skene(model)
}
\arguments{
\item{model}{object of type model}
}
\description{
Check if model is of type Hierarchical Dawid and skene
}
