% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall.R
\name{kendall}
\alias{kendall}
\title{Return a variant of the Kendall tau-a correlation for two rankings}
\usage{
kendall(x, y, unranked_items = NULL)
}
\arguments{
\item{x, y}{ranking vectors of comparable items}

\item{unranked_items}{vector of items that might not be in `x` or}
}
\value{
Numeric value the is the modified Kendall tau rank correlation
}
\description{
The Kendall τ-a correlation is a measure of similarity between two rankings
of the same items.  This function is consistent with this definition, as well
as a modification for rankings with some distinct items.
}
