% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_stan.R
\name{lm_stan}
\alias{lm_stan}
\title{Bayesian linear regression with Stan}
\usage{
lm_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian linear regression with Stan
}
