% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{loo.rater_fit}
\alias{loo.rater_fit}
\alias{loo}
\title{Compute the PSIS LOO CV - a measure of model fit - of a rater fit object.}
\usage{
\method{loo}{rater_fit}(x, ..., cores = getOption("mc.cores", 1))
}
\arguments{
\item{x}{A \code{rater_fit} object. All model types are currently supported
except the basic Dawid-Skene model fit with grouped data.}

\item{...}{Other arguments passed.}

\item{cores}{The number of cores to use when calling the underlying
functions. By default the value of the \code{mc.cores} option.}
}
\value{
A loo object.
}
\description{
Compute the PSIS LOO CV - a measure of model fit - of a rater fit object.
}
\details{
This function is somewhat experimental; model comparison is always
difficult and choosing between variants of the Dawid-Skene model should
be largely guided by considerations of data size and what is known about
the characteristics of the raters. loo is, however, one of the leading
methods for Bayesian model comparison and should provide a helpful guide
in many situations.

When calculating loo we always use the relative effective
sample size, calculated using \code{loo::relaive_eff} to improve the estimates
of the PSIS effective sample sizes and Monte Carlo error.

For further information about the details of loo and PSIS please consult
the provided references.
}
\examples{

\donttest{
fit_ds <- rater(anesthesia, "dawid_skene", verbose = FALSE, chains = 1)
fit_ccds <- rater(anesthesia, "class_conditional_dawid_skene",
                  verbose = FALSE, chains = 1)

loo_ds <- loo(fit_ds)
loo_ccds <- loo(fit_ccds)

# To compare the loos easily we can use the loo_compare function from the
# loo package:
library(loo)

loo_compare(loo_ds, loo_ccds)

# The documentation of the loo package contains more information about how
# the output should be interpreted.
}

}
\references{
Vehtari, A., Gelman, A., and Gabry, J. (2017a). Practical Bayesian model
evaluation using leave-one-out cross-validation and WAIC.
\emph{Statistics and Computing}. 27(5), 1413--1432. doi:10.1007/s11222-016-9696-4
(\href{https://link.springer.com/article/10.1007/s11222-016-9696-4}{journal version},
\href{https://arxiv.org/abs/1507.04544}{preprint arXiv:1507.04544}).

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry, J. (2019).
Pareto smoothed importance sampling.
\href{https://arxiv.org/abs/1507.02646}{preprint arXiv:1507.02646}
}
