% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_complete_rating_design_sim_data}
\alias{make_complete_rating_design_sim_data}
\title{Produce simulation data from a 'complete' rating design}
\usage{
make_complete_rating_design_sim_data(I, J, N)
}
\arguments{
\item{I}{The number of items.}

\item{J}{The number of raters.}

\item{N}{The number of times each rater rates each item.}
}
\value{
Simulation data in the format required by
\code{\link[=simulate_dawid_skene_model]{simulate_dawid_skene_model()}} or \code{\link[=simulate_hier_dawid_skene_model]{simulate_hier_dawid_skene_model()}}.
}
\description{
Produce simulation data from a 'complete' rating design
}
\details{
A 'complete' rating design is situation where every rater rates
each item the same number of times. In this function the number of times
each rater rates each item is \code{N}.
}
\examples{

make_complete_rating_design_sim_data(100, 5, 2)

}
