% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_theta}
\alias{make_theta}
\title{Make a theta parameter}
\usage{
make_theta(diag_values, J, K)
}
\arguments{
\item{diag_values}{The diagonal entries of each error matrix.}

\item{J}{The number of raters (The umber matrices in 3D array).}

\item{K}{The number of latent classes.}
}
\value{
A c(J, K, K) array; the theta parameter
}
\description{
Make a theta parameter
}
\details{
The \code{diag_values} argument can either be a numeric vector of length
1 or J. If it is length J, the jth element is the diagonal values of the
error matrix for the jth rater. If it is length 1 all raters have the same
diagonal values.
}
\examples{

theta <- make_theta(0.7, 5, 4)
theta[1, , ]

}
