% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{markov}
\alias{markov}
\title{Compute ratings from a stochastic matrix}
\usage{
markov(stoc_mtx, beta = 1)
}
\arguments{
\item{stoc_mtx}{a matrix containing the transition information.  The matrix
entry \eqn{A[i,j]} encodes the transistion probability from candidate \eqn{i}
to candidate \eqn{j}}

\item{beta}{a numeric parameter to guarantee an irreducible Markov matrix}
}
\value{
A vector containing the Markov ratings for the candidates
}
\description{
This function computes ratings from a stochastic matrix.
The stochastic matrix can be made irreducible or slightly perturbed by
choosing \eqn{0 < beta < 1}.
}
