% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massey.R
\name{massey}
\alias{massey}
\title{Compute ratings from Massey matrix and net point differentials}
\usage{
massey(massey_mtx, point_diffs)
}
\arguments{
\item{massey_mtx}{a matrix containing the data on number of games played by
each team, against other teams}

\item{point_diffs}{a vector containing the net point differentials for each
team in the games played}
}
\value{
\code{massey_ratings} returns a vector containing the Massey ratings
for each team.
}
\description{
This function computes the Massey ratings for a league from a Massey matrix
and the point differentials for each team.
}
