% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massey.R
\name{massey_adv}
\alias{massey_adv}
\title{Compute the offensive, defensive, and total ratings for each team}
\usage{
massey_adv(massey_mtx, points_for, points_against)
}
\arguments{
\item{massey_mtx}{a matrix containing the data on number of games played
by each team, against other teams}

\item{points_for}{a vector containing the points scored by each team in the
games played}

\item{points_against}{a vector containing the points scored against each team
in the games played}
}
\value{
\code{massey_adv} returns a vector containing the offensive,
defensive, and total Massey ratings for each team.
}
\description{
This function computes the Massey offensive and defensive ratings for each
team.  The total rating, which is just the sum of the offensive and
defensive rating, is also included.
}
