% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Fit noisy catergrical rating models using HMC via Stan}
\usage{
mcmc(data, model, ...)
}
\arguments{
\item{data}{Data for the model in long format as a three column dataframe of
ii, jj, yy}

\item{model}{Model to fit to data of class model}

\item{...}{extra parameters to be passed to the Stan fitting interface}
}
\value{
An object of type fit containing the fitted parameters
}
\description{
Fit noisy catergrical rating models using HMC via Stan
}
