% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\alias{dawid_skene}
\alias{hier_dawid_skene}
\alias{class_conditional_dawid_skene}
\title{Probabilistic models of repeated categorical rating}
\usage{
dawid_skene(alpha = NULL, beta = NULL)

hier_dawid_skene(alpha = NULL)

class_conditional_dawid_skene(alpha = NULL, beta_1 = NULL, beta_2 = NULL)
}
\arguments{
\item{alpha}{prior parameter for pi}

\item{beta}{prior parameter for theta. This can either be a K * K matrix, in
which case it is interpreted as the prior parameter of all of the J
raters, or a J by K by K array in which case it is the fully specified
prior parameter for all raters. (Here K is the number of categories in the
data and J is the number of raters in the data.)}

\item{beta_1}{First on diagonal prior probability parameter}

\item{beta_2}{Second on diagonal prior probability parameter for theta}
}
\value{
a rater model object that can be passed to \code{\link[=rater]{rater()}}.
}
\description{
Functions to set up models and change their prior
parameters for use in \code{\link[=rater]{rater()}}.
}
