% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{multinomial_inits}
\alias{multinomial_inits}
\title{Creates inits for the multinomial model}
\usage{
multinomial_inits(K, J)
}
\arguments{
\item{K}{number of categories}

\item{J}{number of raters}
}
\value{
inits in the format required by stan
}
\description{
Creates inits for the multinomial model
}
