% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{new_long_data}
\alias{new_long_data}
\title{Internal helper to creat a wide_data object}
\usage{
new_long_data(data)
}
\arguments{
\item{data}{a numeric matrix with three columns}
}
\description{
Internal helper to creat a wide_data object
}
