% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{new_mcmc_fit}
\alias{new_mcmc_fit}
\title{make a mcmc rater fit object}
\usage{
new_mcmc_fit(model, draws, data)
}
\arguments{
\item{model}{a rater model}

\item{draws}{a stanfit object}

\item{data}{the data used to fit the model}
}
\value{
a rater fit object
}
\description{
make a mcmc rater fit object
}
