% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{new_table_data}
\alias{new_table_data}
\title{Internal helper to creat a table_data object}
\usage{
new_table_data(data)
}
\arguments{
\item{data}{a numeric matrix}
}
\description{
Internal helper to creat a table_data object
}
