% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{parse_data}
\alias{parse_data}
\title{Convert passed data into Stan data format}
\usage{
parse_data(model, data)
}
\arguments{
\item{data}{data in 'long format'}
}
\value{
Data in format required by Stan
}
\description{
Convert passed data into Stan data format
}
