% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{parse_data_raters}
\alias{parse_data_raters}
\title{Convert passed data into Stan data format for a model wit raters}
\usage{
parse_data_raters(data)
}
\arguments{
\item{data}{data in 'long format'}
}
\value{
Data in format required by Stan
}
\description{
Convert passed data into Stan data format for a model wit raters
}
