% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{parse_priors}
\alias{parse_priors}
\title{Converts default prior parameter specification to full priors}
\usage{
parse_priors(model, K)
}
\arguments{
\item{model}{the rater_model}

\item{K}{the number of categories}
}
\value{
the fully reliased prior parameters
}
\description{
Converts default prior parameter specification to full priors
}
