% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{parse_priors_ds}
\alias{parse_priors_ds}
\title{Converts the passed priors for the Dawid and Skene model}
\usage{
parse_priors_ds(model, data_list)
}
\arguments{
\item{model}{rater model}

\item{data_list}{data converted to Stan list format see \code{parse_data}}
}
\description{
See parse_priors for details
}
