% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{pi_point_estimate}
\alias{pi_point_estimate}
\title{Extract a point esimate of the pi parameter from an MCMC fit}
\usage{
pi_point_estimate(fit, ...)
}
\arguments{
\item{fit}{A rater_fit object}
}
\value{
A vector of length K containing the posterior mean of pi
}
\description{
Extract a point esimate of the pi parameter from an MCMC fit
}
