% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_class_probabilities.rater_fit}
\alias{plot_class_probabilities.rater_fit}
\title{Plot the latent class estimates of a rater_fit object}
\usage{
\method{plot_class_probabilities}{rater_fit}(x, ...)
}
\arguments{
\item{x}{rater fit object}

\item{...}{Other arguments}
}
\value{
Plot of the rate accuracy estimates
}
\description{
Plot the latent class estimates of a rater_fit object
}
