% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_class_probabilities}
\alias{plot_class_probabilities}
\title{Generic to plot the latent class probabilites}
\usage{
plot_class_probabilities(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{extra stuff}
}
\description{
Generic to plot the latent class probabilites
}
\details{
This method can either be used to plot matrices or rater_fit
  objects
}
