% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_latent_class}
\alias{plot_latent_class}
\title{Plot the latent class estimates}
\usage{
plot_latent_class(fit)
}
\arguments{
\item{fit}{rateR fit object}
}
\value{
Plot of the rate accuracy estimates
}
\description{
Plot the latent class estimates
}
