% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi_dist}
\alias{plot_pi_dist}
\title{Plot the posterior distribution of prevalence probabilities}
\usage{
plot_pi_dist(fit)
}
\arguments{
\item{fit}{A rater fit object}
}
\value{
A plot of posterior distributions of prevalence probabilities of latent classes
}
\description{
Plot the posterior distribution of prevalence probabilities
}
