% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_prevalance}
\alias{plot_prevalance}
\title{Plot the popuation prevalence estimates}
\usage{
plot_prevalance(fit)
}
\arguments{
\item{fit}{rater fit object}
}
\value{
Plot of the population prevelance esitmates extracted in fits
}
\description{
Plot the popuation prevalence estimates
}
