% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_raters}
\alias{plot_raters}
\title{Plot the rater accuracy estimates}
\usage{
plot_raters(fit)
}
\arguments{
\item{fit}{rateR fit object}
}
\value{
Plot of the rate accuracy estimates
}
\description{
Plot the rater accuracy estimates
}
