% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_theta}
\alias{plot_theta}
\title{Plot the rater accuracy estimates}
\usage{
plot_theta(fit, which = NULL)
}
\arguments{
\item{fit}{rater fit object}

\item{which}{which raters to plot}
}
\value{
Plot of the rate accuracy estimates
}
\description{
Plot the rater accuracy estimates
}
