% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_z.matrix}
\alias{plot_z.matrix}
\title{Plot the latent class estimates of a matrix}
\usage{
\method{plot_z}{matrix}(x, ...)
}
\arguments{
\item{x}{numeric matrix object}
}
\value{
Plot of the rate accuracy estimates
}
\description{
Plot the latent class estimates of a matrix
}
