% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_z.rater_fit}
\alias{plot_z.rater_fit}
\title{Plot the latent class estimates of a rater_fit object}
\usage{
\method{plot_z}{rater_fit}(x, ...)
}
\arguments{
\item{x}{rater fit object}
}
\value{
Plot of the rate accuracy estimates
}
\description{
Plot the latent class estimates of a rater_fit object
}
