% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_z}
\alias{plot_z}
\title{Plot the latent class estimates}
\usage{
plot_z(fit)
}
\arguments{
\item{fit}{rater fit object}
}
\value{
Plot of the rate accuracy estimates
}
\description{
Plot the latent class estimates
}
