% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{point_estimate}
\alias{point_estimate}
\title{Extract point estimates of parameters from a fit object}
\usage{
point_estimate(fit, pars = c("pi", "theta", "z"), ...)
}
\arguments{
\item{fit}{A rater_fit object}

\item{pars}{A character vector of parmeter names to select}

\item{format}{TODO}
}
\value{
TODO
}
\description{
Extract point estimates of parameters from a fit object
}
\details{
If the passed fit object was fit using MCMC then the posterior
  means are used. If it was fit through optimisation the MAP esimates
  are returned.
}
