% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{posterior_draws}
\alias{posterior_draws}
\title{Extract posterior draws from a rater_fit object}
\usage{
posterior_draws(fit, pars = c("pi", "theta"), ...)
}
\arguments{
\item{fit}{A rater_fit object}

\item{pars}{A specification of which parameters to return draws from}

\item{...}{Extra arguments}
}
\value{
If multiple draws are requested then a list of arrays, otherwise
 a single array.
}
\description{
Extract posterior draws from a rater_fit object
}
