% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{posterior_interval.mcmc_fit}
\alias{posterior_interval.mcmc_fit}
\title{Extract posterior intervals for parameters of the model}
\usage{
\method{posterior_interval}{mcmc_fit}(object, prob = 0.9, pars = c("pi", "theta"), ...)
}
\arguments{
\item{pars}{}
}
\description{
Extract posterior intervals for parameters of the model
}
