% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Extract posterior draws from a rater_fit object}
\usage{
posterior_samples(fit, pars = c("pi", "theta"), ...)
}
\arguments{
\item{fit}{A rater_fit object}

\item{pars}{A specification of which parameters to return draws from}

\item{...}{Extra arguments}
}
\value{
A named list of parameter draws
}
\description{
Extract posterior draws from a rater_fit object
}
