% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_class.R
\name{print.fit}
\alias{print.fit}
\title{print a fit object}
\usage{
\method{print}{fit}(x, ...)
}
\arguments{
\item{x}{fit object to be printed}

\item{...}{other args passed to the function}
}
\description{
print a fit object
}
