% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{print.mcmc_fit}
\alias{print.mcmc_fit}
\title{Print a mcmc_fit object}
\usage{
\method{print}{mcmc_fit}(x, ...)
}
\arguments{
\item{x}{fit object to be printed}

\item{...}{other args passed to the function}
}
\description{
Print a mcmc_fit object
}
