% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{print.model}
\alias{print.model}
\title{print a fit model object}
\usage{
\method{print}{model}(x, ...)
}
\arguments{
\item{x}{fit object to be printed}

\item{...}{other args passed to the function}
}
\description{
print a fit model object
}
