% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{print.optim_fit}
\alias{print.optim_fit}
\title{Print a optim_fit object}
\usage{
\method{print}{optim_fit}(x, ...)
}
\arguments{
\item{x}{fit object to be printed}

\item{...}{other args passed to the function}
}
\description{
Print a optim_fit object
}
