% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_model_class.R
\name{print.rater_model}
\alias{print.rater_model}
\title{print a rater_model object}
\usage{
\method{print}{rater_model}(x, ...)
}
\arguments{
\item{x}{a rater_model object}

\item{...}{other args}
}
\description{
print a rater_model object
}
