% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{prior_summary.rater_fit}
\alias{prior_summary.rater_fit}
\alias{prior_summary}
\title{Provide a summary of the priors specified in a \code{rater_fit} object.}
\usage{
\method{prior_summary}{rater_fit}(object, ...)
}
\arguments{
\item{object}{A \code{rater_fit} object.}

\item{...}{Other arguments.}
}
\description{
Provide a summary of the priors specified in a \code{rater_fit} object.
}
\examples{
\donttest{
# Fit a model using MCMC (the default).
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

# Summarise the priors (and model) specified in the fit.
prior_summary(fit)

}

}
