% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankplot.R
\name{rankgraph}
\alias{rankgraph}
\title{Create the graph object for two rankings}
\usage{
rankgraph(x, y, other_rankings = NULL)
}
\arguments{
\item{x, y}{ranking vectors of comparable items}

\item{other_rankings}{vector of items that might not be in `x` or `y`}
}
\description{
This returns a graph object in the sense of the \code{DiagrammeR} package.
This allows for some customization prior to rendering.  You can read the
documentation here: \url{http://rich-iannone.github.io/DiagrammeR/graph_creation.html}.
}
