% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankplot.R
\name{rankplot}
\alias{rankplot}
\title{Render the graph for two rankings}
\usage{
rankplot(x, y, other_rankings = NULL)
}
\arguments{
\item{x, y}{ranking vectors of comparable items}

\item{other_rankings}{vector of items that might not be in `x` or `y`}
}
\description{
Render the graph for two rankings
}
