% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateR-package.R
\docType{package}
\name{rateR-package}
\alias{rateR-package}
\alias{rateR}
\title{The 'rateR' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
