% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{loo_compare}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{loo}{\code{\link[loo]{loo_compare}}}
}}

