% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_model}
\alias{simulate_model}
\title{Simulate from a rater model}
\usage{
simulate_model(model, parameters, data_list, ...)
}
\arguments{
\item{...}{}

\item{object}{a rater model object}

\item{nsim}{the number of simulation to perform}

\item{seed}{the seed for the simulation}

\item{new_params}{new parameters for the model}
}
\value{
simulation
}
\description{
Simulate from a rater model
}
\details{
if new params is specifed data will be generated from the model
  specifed by a model is new_params parameters
}
\examples{




}
