% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{summary.model}
\alias{summary.model}
\title{Summary of model}
\usage{
\method{summary}{model}(object, ...)
}
\arguments{
\item{object}{model object}

\item{...}{other args passed to the function}
}
\description{
Summary of model
}
