% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{summary.optim_fit}
\alias{summary.optim_fit}
\title{Summary of optim fit}
\usage{
\method{summary}{optim_fit}(object, ...)
}
\arguments{
\item{object}{object of type rater fit}

\item{...}{other args passed to function}
}
\description{
Summary of optim fit
}
