% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_model_class.R
\name{summary.rater_model}
\alias{summary.rater_model}
\title{summarise a rater_model}
\usage{
\method{summary}{rater_model}(object, ...)
}
\arguments{
\item{object}{a rater_model object}

\item{...}{other args}
}
\description{
summarise a rater_model
}
