% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{theta_point_estimate}
\alias{theta_point_estimate}
\title{Extract rater accuracy estimates for the Dawid Skene models}
\usage{
theta_point_estimate(fit, which = NULL, ...)
}
\arguments{
\item{fit}{fit object}

\item{which}{which raters to extract}

\item{...}{extra args}
}
\value{
list of matrices containing probability confusion matrices for each
 rater
}
\description{
Extract rater accuracy/theta estimates from a Dawid Skene fit object
}
