% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankplot.R
\name{unranked}
\alias{unranked}
\title{Get the unranked items for each list in the comparison}
\usage{
unranked(x, y, other_rankings = NULL)
}
\description{
Get the unranked items for each list in the comparison
}
