% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{validate_alpha}
\alias{validate_alpha}
\title{Checks that a possible prior on alpha is valid}
\usage{
validate_alpha(alpha, K)
}
\arguments{
\item{alpha}{prior parameter for pi}

\item{K}{number of categories in the data}
}
\description{
Checks that the length of the prior on alpha == K and errors if not
}
