% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{validate_data}
\alias{validate_data}
\title{Check that the passed data is in the appropriate format}
\usage{
validate_data(model, data)
}
\arguments{
\item{data}{Data passed to \code{mcmc} hopefully in 'long' format}
}
\description{
Checks whether the passed data is in the long format. That is: three columns
item index, annotator index, annotation. This function checks that the data
has three columns and that they are all numeric.
}
