% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_fit}
\alias{validate_fit}
\title{Check that a passed object is actually of type fit}
\usage{
validate_fit(fit)
}
\arguments{
\item{fit}{fit object}
}
\description{
Check that a passed object is actually of type fit
}
