% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{validate_input}
\alias{validate_input}
\title{Helper to check if passed data and model are valid and consistent}
\usage{
validate_input(data, model)
}
\arguments{
\item{data}{rater_data}

\item{model}{rater_model}
}
\description{
Helper to check if passed data and model are valid and consistent
}
