% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate passed parameters}
\usage{
validate_parameters(pars)
}
\arguments{
\item{parameters}{a (named) list of parmeters}
}
\description{
Checks the parameter are of the appropritate type/form and that they are
self consistent
}
