% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_data.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Coerce from wide to long format}
\usage{
wide_to_long(data)
}
\arguments{
\item{data}{in wide format}
}
\value{
Data in long format
}
\description{
Coerces data in a wide format (possibly with NA's) to a long format
}
