% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{z_point_estimate}
\alias{z_point_estimate}
\title{Extract latent class estimates from a fit}
\usage{
z_point_estimate(fit, ...)
}
\arguments{
\item{fit}{fit object}

\item{...}{extra args}
}
\value{
Latent class estimates: A vector length I consisting of whole
numbers from 1 to K.
}
\description{
Extract latent class estimates from a fit
}
\details{
This function returns actual estimates of the latent class i.e.
whole numbers from 1 to K. These are found by taking the class with the
highest probability.
}
