% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall.R
\name{zipit}
\alias{zipit}
\title{Auxiliary function, not for export
Returns a list of 2-vectors, [(x1,y1),...,(xn,yn)]}
\usage{
zipit(x, y)
}
\description{
Auxiliary function, not for export
Returns a list of 2-vectors, [(x1,y1),...,(xn,yn)]
}
