% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anesthesia}
\alias{anesthesia}
\title{Anaesthetist ratings for patient suitability for surgery}
\format{
A matrix with 315 rows and 3 columns:
\describe{
\item{item}{item index, which item is being rated}
\item{rater}{rater index, which rater is doing the rater}
\item{rating}{what rating was given}
}
}
\usage{
anesthesia
}
\description{
A dataset containing anaesthetist ratings of patient readiness for surgery for
5 anaesthetists on 45 patients with 4 categories in 'long' format. Each row
represents one unique rating.
}
\keyword{datasets}
