% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caries}
\alias{caries}
\title{Caries ratings of whether caries are 'sound' or not (c...)}
\format{
An object of class \code{data.frame} with 32 rows and 6 columns.
}
\usage{
caries
}
\description{
TODO
}
\keyword{datasets}
