% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{creat_inits}
\alias{creat_inits}
\title{Creates inits for the stan MCMC chains}
\usage{
creat_inits(model)
}
\arguments{
\item{model}{}
}
\description{
Creates inits for the stan MCMC chains
}
\details{
If the model is of type Dawid and Skene creates inits favouring
good raters. This is to prevent label switching see: TODO. If the model is
not Dawid and Skene then the inits are set randomly
}
