% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{data_types}
\alias{data_types}
\alias{wide_data}
\alias{long_data}
\alias{grouped_data}
\title{Data types compatible with the models of categorical annotation}
\usage{
wide_data(data)

long_data(data)

grouped_data(data)
}
\arguments{
\item{data}{a numeric matrix}
}
\value{
a \code{rater_data} object that can be passed to \code{\link{rater}}.
}
\description{
Functions to coerce/parse various types of input data
  for use in \code{\link{rater}()}.
}
