% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_class.R
\name{extract.fit}
\alias{extract.fit}
\title{Extract method for a rater fit object}
\usage{
\method{extract}{fit}(x, ...)
}
\arguments{
\item{fit}{a fit object}

\item{type}{what sort of plot should be plotted}

\item{which}{Which of the raters' matrices to extract (only used for type =
theta or raters)}
}
\description{
Extract different paramter estimates from a fit object
}
