% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_class.R
\name{extract.fit}
\alias{extract.fit}
\title{Extract method for a rater fit object}
\usage{
\method{extract}{fit}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{other arguments to be passed
theta or raters)}
}
\description{
Extract different paramter estimates from a fit object
}
\details{
... must contain the param argument which tell the function
what to extract from the fit object. It may also contain the which argument
which controls which raters confusion matrices will be plotted.
}
