% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_latent_class}
\alias{extract_latent_class}
\title{Extract latent class estimates from a fit}
\usage{
extract_latent_class(fit)
}
\arguments{
\item{rater}{fit object}
}
\value{
Probalistic latent class measurements
}
\description{
Extract latent class estimates from a fit
}
