% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_pi.mcmc_fit}
\alias{extract_pi.mcmc_fit}
\title{Extract prevalence information}
\usage{
\method{extract_pi}{mcmc_fit}(fit, ...)
}
\arguments{
\item{fit}{fit object}

\item{...}{extra args}
}
\description{
Extract prevalence/pi estiamtes from a fit object
}
