% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{extract_raters}
\alias{extract_raters}
\title{Function to extract raters from a fit}
\usage{
extract_raters(fit, which = NULL, ...)
}
\arguments{
\item{which}{which rater to select}

\item{...}{extra args}

\item{x}{an object}
}
\description{
Function to extract raters from a fit
}
