% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_raters}
\alias{extract_raters}
\title{Extract rater accuracy estimates}
\usage{
extract_raters(fit, which = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
list of matrices containing probability confusion matrices for each
 rater
}
\description{
Extract rater accuracy/theta estimates from a fit object
}
