% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_raters_multi}
\alias{extract_raters_multi}
\title{Extract rater accuarcy estimates}
\usage{
extract_raters_multi(fit)
}
\arguments{
\item{fit}{fit object}
}
\value{
list of matrices containing proabaility confusion matrices for each
 rater
}
\description{
Extract ater accuarcy/theta estimates from a fit object
}
