% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{extract_z}
\alias{extract_z}
\title{Generic to extract z (latent class) from a fit}
\usage{
extract_z(fit, ...)
}
\arguments{
\item{...}{extra stuff}

\item{x}{an object}
}
\description{
Generic to extract z (latent class) from a fit
}
