% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{mcmc_diagnostics}
\alias{mcmc_diagnostics}
\title{Retrieve MCMC convergence diagnostics for a rater fit}
\usage{
mcmc_diagnostics(fit, pars = c("pi", "theta"))
}
\arguments{
\item{fit}{an MCMC rater_fit object}

\item{pars}{the parameters to return the diagnostics for}
}
\value{
A matrix where the columns represent different diagnostics and the
  rows are different parameters.
}
\description{
Retrieve MCMC convergence diagnostics for a rater fit
}
