% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\alias{dawid_skene}
\alias{hier_dawid_skene}
\alias{class_conditional_dawid_skene}
\title{Probablistic models of categorical data annotation}
\usage{
dawid_skene(alpha = NULL, beta = NULL)

hier_dawid_skene(alpha = NULL)

class_conditional_dawid_skene(alpha = NULL, beta_1 = NULL, beta_2 = NULL)
}
\arguments{
\item{alpha}{prior parameter for pi}

\item{beta}{prior parameter for theta}
}
\value{
a \code{model} object that can be passed to \code{\link{rater}}.
}
\description{
Functions to set up models and change their prior
  parameters for use in \code{\link{rater}()}.
}
