% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{new_optim_fit}
\alias{new_optim_fit}
\title{Make a optim_rater fit object}
\usage{
new_optim_fit(model, estimates, data)
}
\arguments{
\item{model}{a rater model}

\item{estimates}{a stan optimisation object}

\item{data}{the data used to fit the model}
}
\value{
a rater fit object
}
\description{
Make a optim_rater fit object
}
