% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{parse_priors}
\alias{parse_priors}
\title{Converts the passed priors into the appropriate format for Stan}
\usage{
parse_priors(model, data_list)
}
\arguments{
\item{model}{the model passed to mcmc}

\item{data_list}{data already parsed into list format}
}
\description{
In paticular this function creates default priors if no priors are passed
and checks if passed priors are of the correct dimension.
}
\details{
This function needs to see the passed data to check that the priors
are the correct dimension. It needs to see the model type because diffrent
models require different priors. This function dispatches into two sub
functions depending on which model is passed
}
