% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{pi_point_estimate}
\alias{pi_point_estimate}
\alias{pi_point_estimate.mcmc_fit}
\alias{pi_point_estimate.optim_fit}
\title{Extract a point estimate of the pi parameter from an MCMC fit}
\usage{
pi_point_estimate(fit, ...)

\method{pi_point_estimate}{mcmc_fit}(fit, ...)

\method{pi_point_estimate}{optim_fit}(fit, ...)
}
\arguments{
\item{fit}{A rater_fit object}

\item{...}{Other arguments}
}
\value{
A vector of length K containing the posterior mean of pi
}
\description{
Extract a point estimate of the pi parameter from an MCMC fit
}
