% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_class.R
\name{plot.fit}
\alias{plot.fit}
\title{Plot a fit object}
\usage{
\method{plot}{fit}(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{other arguments to be passed
theta or raters)}
}
\description{
Plot a fit object
}
\details{
... must contain the param argument which tell the function
what to extract from the fit object. It may also contain the which argument
which controls which raters confusion matrices will be plotted.
}
