% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_class.R
\name{plot.fit}
\alias{plot.fit}
\title{Plot a fit object}
\usage{
\method{plot}{fit}(x, ...)
}
\arguments{
\item{fit}{a fit object}

\item{type}{what sort of plot should be plotted}

\item{which}{Which of the raters' matrices to extract (only used for type =
theta or raters)}
}
\description{
Plot a fit object
}
