% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{plot.rater_fit}
\alias{plot.rater_fit}
\title{Plot a rater_fit object}
\usage{
plot.rater_fit(x, type = "theta", ...)
}
\arguments{
\item{x}{fit object}

\item{type}{the type of plot}

\item{...}{other args}
}
\description{
Plot a rater_fit object
}
\details{
... must contain the param argument which tell the function
what to extract from the fit object. It may also contain the which argument
which controls which raters confusion matrices will be plotted.
}
