% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{plot.rater_fit}
\alias{plot.rater_fit}
\title{Plot a rater_fit object}
\usage{
\method{plot}{rater_fit}(x, pars = c("pi", "theta", "z"), ...)
}
\arguments{
\item{x}{A rater_fit object}

\item{pars}{Which parameters to plot. Can use both mathematical or
natural names.}

\item{...}{Other arguments. This should contain the which argument for
theta plots.}
}
\value{
If one parameter is requested a ggplot2 plot. If multiple parameters
  are requested a list of ggplot2 plots.
}
\description{
Plot a rater_fit object
}
