% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pi}
\alias{plot_pi}
\title{Plot the population prevalence estimates}
\usage{
plot_pi(fit)
}
\arguments{
\item{fit}{rater fit object}
}
\value{
Plot of the population prevalence estimates extracted in fits
}
\description{
Plot the population prevalence estimates
}
