% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_z}
\alias{plot_z}
\title{Generic to plot z (latent class)}
\usage{
plot_z(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{extra stuff}
}
\description{
Generic to plot z (latent class)
}
\details{
This method can either be used to plot matrices or rater_fit
  objects
}
