% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.R
\name{posterior_interval.optim_fit}
\alias{posterior_interval.optim_fit}
\title{Extract posterior intervals for parameters of the model}
\usage{
\method{posterior_interval}{optim_fit}(object, prob = 0.9, pars = c("pi", "theta"), ...)
}
\arguments{
\item{object}{A rater optim_fit object}

\item{prob}{A probability}

\item{pars}{The parameters to calculate the intervals for}

\item{...}{Other arguments}
}
\description{
Extract posterior intervals for parameters of the model
}
