% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{print.mcmc_fit}
\alias{print.mcmc_fit}
\title{Print a \code{mcmc_fit} object}
\usage{
\method{print}{mcmc_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcmc_fit}.}

\item{...}{Other arguments.}
}
\description{
Print a \code{mcmc_fit} object
}
\examples{
\donttest{

# Suppress sampling output.
mcmc_fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)
print(mcmc_fit)

}

}
