% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{print.optim_fit}
\alias{print.optim_fit}
\title{Print a \code{optim_fit} object}
\usage{
\method{print}{optim_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{optim_fit}.}

\item{...}{Other arguments.}
}
\description{
Print a \code{optim_fit} object
}
\examples{
\donttest{

optim_fit <- rater(anesthesia, "dawid_skene", method = "optim")
print(optim_fit)

}

}
