% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_model_class.R
\name{print.rater_model}
\alias{print.rater_model}
\title{Print a \code{rater_model} object.}
\usage{
\method{print}{rater_model}(x, ...)
}
\arguments{
\item{x}{A \code{rater_model} object.}

\item{...}{Other arguments}
}
\description{
Print a \code{rater_model} object.
}
\examples{
mod <- dawid_skene()
print(mod)

}
