% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{rater}
\alias{rater}
\title{Fit noisy catergrical rating models using Stan}
\usage{
rater(data, model, method = "mcmc", inits = NULL, ...)
}
\arguments{
\item{data}{Data for the model in long format as a three column dataframe of
ii, jj, yy}

\item{model}{Model to fit to data of class model}

\item{method}{method the method used to fit the model}

\item{inits}{the initialization points of the fitting algorithm}

\item{...}{extra parameters to be passed to the Stan fitting interface}
}
\value{
An object of type fit containing the fitted parameters
}
\description{
Fit noisy catergrical rating models using Stan
}
