% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{rater}
\alias{rater}
\title{Fit statistical models of noisy categorical rating data using Stan}
\usage{
rater(data, model, method = "mcmc", data_format = "long", inits = NULL, ...)
}
\arguments{
\item{data}{A 2D data object: data.frame, matrix, tibble etc. with data in
either long or grouped format.}

\item{model}{Model to fit to data - must be rater_model or a character
string - the name of the model. If the character string is used, the
prior parameters will be set to their default values.}

\item{method}{A length 1 character vector, either "mcmc" or "optim". This
represents the fitting method used by Stan.}

\item{data_format}{A length 1 character vector, either "long" or "grouped".
The format that the passed data is in. Defaults to "long".}

\item{inits}{The initialization points of the fitting algorithm}

\item{...}{Extra parameters which are passed to the Stan fitting interface}
}
\value{
An object of type class rater_fit containing the fitted parameters.
}
\description{
This functions allows the user to fit statistical models of noisy
categorical rating, based on the Dawid-Skene model, using Bayesian
inference. A variety of data formats and models are supported. Inference
is done using Stan, allowing models to be fit efficiently, using both
optimisation and Markov Chain Monte Carlo (MCMC).
}
\details{
The MCMC algorithm used by Stan is No U Turn Sampling.
}
\examples{
# Fit a model using optimisation.
optim_fit <- rater(anesthesia, dawid_skene(), method = "optim")

# Fit a model using grouped data (and optimisation).
grouped_fit <- rater(caries, dawid_skene(), data_format = "grouped",
                     method = "optim")

}
