% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateR.R
\name{rateR}
\alias{rateR}
\title{A shiny application for}
\usage{
rateR()
}
\description{
rateR provides an intuitive interface with options to
}
\examples{
if (interactive()) {
  rateR::rateR()
}
}
\seealso{
Web version: \url{https://lmarusich.shinyapps.io/rateR/}
}
