% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_ds}
\alias{simulate_ds}
\title{Internal funtion to simulate the dawid and skene model}
\usage{
simulate_ds(model, parameters, data_list, ...)
}
\arguments{
\item{...}{other stuff}

\item{object}{a model object}
}
\value{
simulated data
}
\description{
Internal funtion to simulate the dawid and skene model
}
\details{
Some notes:
  * Currently this does not support direct passing of z
  * There is basically no input checking yet - due the high number of inputs
  * The slightly clunky interface for new model simulation is due to needing
    to use the stats::simulate generic
}
