% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_fit_class.R
\name{summary.mcmc_fit}
\alias{summary.mcmc_fit}
\title{Summary of fit}
\usage{
\method{summary}{mcmc_fit}(object, ...)
}
\arguments{
\item{object}{object of type rater fit}

\item{...}{other args passed to function}
}
\description{
Summary of fit
}
