% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater_model_class.R
\name{summary.rater_model}
\alias{summary.rater_model}
\title{Summarise a \code{rater_model}.}
\usage{
\method{summary}{rater_model}(object, ...)
}
\arguments{
\item{object}{A \code{rater_model} object.}

\item{...}{Other arguments.}
}
\description{
Summarise a \code{rater_model}.
}
\examples{
mod <- dawid_skene()
summary(mod)

}
